게시글 수정기능

+++ src/main/webapp/jsp/article/detail.jsp
@@ -22,7 +22,12 @@
 	<div>내용 : <%= (String) articleRow.get("body") %></div>
 	<div>
 		<a href="list">목록</a>
-		<a href="doDelete?id=<%= (int) articleRow.get("id") %>">삭제</a>
+		<a href="modify?id=<%= (int) articleRow.get("id") %>">수정</a>
+		<a href="doDelete?id=<%= (int) articleRow.get("id") %>" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;">삭제</a>
 	</div>
 </body>
 </html>
+
+
+
+
